+++ marvel/src/components/Card/CardStyle.ts
