+++ marvel/src/components/Card/index.ts
@@ -1,1 +1,1 @@
-export { default as Card } from "./card.component";
+export { default as Card } from "./Card";
