+++ marvel/src/components/NavigationButton/index.ts
@@ -1,0 +1,1 @@
+export { default as NavigationButton } from "./NavigationButton";
