+++ marvel/src/components/Search/index.ts
@@ -1,1 +1,1 @@
-export { default as Search } from "./search.component";
+export { default as Search } from "./Search";
