--- marvel/src/components/navigation-button/index.ts
@@ -1,1 +1,0 @@
-export { default as NavigationButton } from "./navigation-button.component";
