+++ marvel/src/components/ui/footer/index.ts
@@ -1,1 +1,1 @@
-export { default as Footer } from "./footer.component";
+export { default as Footer } from "./Footer";
