+++ marvel/src/components/ui/header/index.ts
@@ -1,1 +1,1 @@
-export { default as Header } from "./header.component";
+export { default as Header } from "./Header";
