+++ marvel/src/components/ui/index.ts
@@ -1,4 +1,4 @@
-import { Header } from "./header";
-import { Footer } from "./footer";
+import { Header } from "./Header";
+import { Footer } from "./Footer";
 
 export { Header, Footer };
