+++ marvel/src/frame/MainFrame/index.ts
@@ -1,0 +1,1 @@
+export { default as MainFrame } from "./MainFrame";
