--- marvel/src/frame/main-frame/index.ts
@@ -1,1 +1,0 @@
-export { default as MainFrame } from "./main-frame.component";
