+++ marvel/src/main.tsx
@@ -1,7 +1,10 @@
 import React from "react";
+
 import ReactDOM from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
-import App from "./App";
+
+// Componerts
+import App from "App";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
